EventFactory implemented
Event running makes more sense, with a single event factory called with
event paramaters to create event archives on the day the event is held
only. Got rid of scheduling events in the future as only the human
player would really benefit from such a feature and we can add in
something to handle it later if necessary.